requirements.txt
asgiref==3.5.2
boto3==1.24.64
botocore==1.27.64
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.1
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==2.8.0
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-storages==1.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
jmespath==1.0.1
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
